[mallard] Finished inline_link and removed inline_quote.
I'm not wholly convinced that an inline quote element serves any
useful purpose.  It never generates any styling effects.  It only
adds punctuation.  Punctuation are just characters that, as far
as I'm concerned, are properly part of the text content.

Granted, it's useful for making fancy quotation marks, but I don't
think it's worth the processing baggage.  It seems simple at first,
but it turns out that the punctuation rules in different languages
are tricky.  I think it's best that people just type the punctuation
they intend to have.