yelp-check.py: Allow custom checkers defined in the config file
This commit allows you to specify your own checkers in .yelp-tools.cfg.
This concept of how these work is very similar to Schematron, but very
simplified in a basic config file syntax. For example, here's how to
write a check to make sure you always have a non-empty desc element:

  [namespaces]
  mal = http://projectmallard.org/1.0/

  [check:gnome-desc]
  select = /mal:page/mal:info
  assert = normalize-space(mal:desc) != ''
  message = Must have non-empty desc