wip! tools: Add a url handler that generates html from documentation
Some systems don't install yelp by default but still want to show
documentation.

This commit adds a small url handler to automatically convert
documentation to html format and then launching a web browser to show
it.

This is work in progress and not ready to go in as-is. A few things
that need work:

1. Currently it only supports untranslated documentation
2. Firefox in a flatpak doesn't quite work yet.
3. It unconditionally generates the documentation every time. It should
instead skip that step if the docs are already generated
4. It's very poorly tested. It needs to be tested more and any issues
fixed.